Improve packets code to send less data

+++ Items/Accessories/Options/Miscellaneous/OptionSeed.cs
@@ -43,7 +43,7 @@ namespace ChensGradiusMod.Items.Accessories.Options.Miscellaneous
           ModOwner(player).seedProjectile = Main.projectile[pInd];
         }
 
-        ModOwner(player).seedRotateDirection = -hideVisual.ToDirectionInt();
+        ModOwner(player).seedRotateDirection = (sbyte)-hideVisual.ToDirectionInt();
       }
     }
